address PR comments: rename, remove zindex, add comment
