+++ src/components/loaders/LoadingSpinner/index.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react'
 import { describe, expect, it } from 'vitest'
 
-import RcSesLoadingSpinner from './LoadingSpinner'
+import RcSesLoadingSpinner from './index'
 
 describe('RcSesLoadingSpinner', () => {
   it('should render a progress bar', () => {
