+++ src/library/index.ts
@@ -46,9 +46,9 @@ import RcSesCardFormContainer from '@/components/layout/ServiceFormContainer/Car
 import RcSesServiceHeader from '@/components/layout/ServiceHeader'
 import RcSesServicePage from '@/components/layout/ServicePage'
 import ServiceWizardStepper from '@/components/layout/ServiceWizardStepper'
-import RcSesFullPageLoader from '@/components/loaders/FullPageLoader/FullPageLoader'
-import RcSesLoader from '@/components/loaders/Loader/Loader'
-import RcSesLoadingSpinner from '@/components/loaders/LoadingSpinner/LoadingSpinner'
+import RcSesFullPageLoader from '@/components/loaders/FullPageLoader'
+import RcSesLoader from '@/components/loaders/Loader'
+import RcSesLoadingSpinner from '@/components/loaders/LoadingSpinner'
 import RcSesDialog from '@/components/overlays/Dialog'
 import RcSesModal from '@/components/overlays/Modal'
 import RcSesTheme from '@/theme/light'
